requirements-dev.txt
pytest>=8.0,<9.0
Pillow>=10.0,<12.0
python-barcode>=0.15,<1.0
